+++ HISTORY.rst
@@ -5,9 +5,10 @@
 History
 =======
 
-current - 2018-08-03 - 0.00Mb
+current - 2018-08-05 - 0.00Mb
 =============================
 
+* `1`: [REMOVED] prepare an example with C# (2018-08-05)
 * `6`: implements an ExtensionArray from pandas with a C++ type (2018-08-03)
 * `4`: implements a series with WeightedDouble (2018-08-03)
 * `3`: add an example based on pybind11 (2018-08-02)